New Development: No             JIRA Issue: No List JIRA ticket.         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: Code cleanup, fix shadowing code
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@36062 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/atnf/atca/ATCAFiller.h
@@ -32,8 +32,6 @@
 #include <casa/aips.h>
 #include <casa/Arrays/Matrix.h>
 #include <casa/Logging/LogIO.h>
-//#include <tables/Tables/ArrayColumn.h>
-//#include <tables/Tables/ScalarColumn.h>
 #include <tables/DataMan/TiledDataStManAccessor.h>
 #include <ms/MeasurementSets/MeasurementSet.h>
 #include <ms/MeasurementSets/MSColumns.h>
@@ -103,7 +101,7 @@ public:
   ATCAFiller & numChan1(Int numchan1);
 
   // Set shadowing limit
-  ATCAFiller & shadow(Double diam);
+  ATCAFiller & shadow(Float diam);
 
   // Deselect antennas.
   ATCAFiller & deselectAntenna(Vector<Int> antennas);
